requirements.txt
requests
pysocks
colorama